[#19] Make `ignored` consider only exact matches
Problem: it's not good that `ignored` matches at file prefix, then if
someone marked `doc` folder as ignored, `doc-others` folder would be
ignored as well.

Solution: use exact matching. This requires user specifying
entries of `ignored` in specific format (e.g. `doc`, not `./doc`), but
let's don't bother about this for now.

Also, we already can ignore not only folders but also files, so we now
mention this in documentation.