[#55] Account for Markdown flavor type
Problem: it turned out that GitHub and GitLab render anchors for headers
differently.

Solution: make it possible to specify flavor in config, adjust
headers conversion respectively. Added some fixes for GitHub as well.

Flavor field is mandatory, so this change is breaking, but I
think it's a worthy thing since otherwise other subtle bugs are
possible, it should be good if users specify the flavor as soon
as possible.