[#72] Fix anchor generation of headers with spaces
Problem: A header name with its custom anchor appended after the
octothorpe symbol(s) is parsed with the leading spaces present, thus
generating an anchor from it will result in having a leading hyphen
prepended to its beginning.

Solution: Strip the header name to ensure the absence of the leading
hyphen in the later generated anchor.