[#49] Provide error messages for illegal square brackets in URIs
Problem: According to RFC 3986 section 2.2, square brackets are reserved
characters and should be percent-encoded in the URIs; however, when
they are used outside their reserved purpose in the respective context,
they MUST be percent-encoded. Currently, xrefcheck does not provide a
verbose error message when such character is detected in the URI.

Solution: Provide a better means to indicate that such a character has
been detected by supplying an error message describing the location
of the said character and its percent-encoded octet.