[#121] Improve help message of `--mode` command line option
Problem: The help message of the `--mode` option doesn't say which modes
are supported. The only way for a user to find that out is to
purposefully call `--mode` with a garbage value, and then xrefcheck will
complain about it and print the list of supported modes.

Solution: Improve the help message to display the list of supported
modes.