[Chore] Update release instructions
Problem: Turns out the release instructions are not entirely accurate.

In order for `xrefcheck-action` to work, the release page must have the
linux binary and the windows zip file attached.

Turns out the recommended way to do this is to edit the "auto-release"
page created by the "Create a pre-release" CI step, see the instructions
here:
https://raw.githubusercontent.com/serokell/scratch/release-binary/scripts/release-binary.sh

Solution: edit the release checklist to include these instructions.