[#162] Remove artificial delay in `allowRewrite`
Problem: `allowRewrite` uses `threadDelay (ms 100)` to "prevent our
output to interleave with further outputs".

It's not exactly what problem this is supposed to address, and how it's
supposed to fix it. This seems like it may have been an earlier attempt
to fix the issue addressed in the previous commit.

Solution: delete `threadDelay`.