[#165] [#192] Add tests for specific cases
Problem: there is no tests for links to directories and non-markdown files.
Also we aren't testing that links are case sensetive

Solution: add such tests to `check-local-refs`, also check that only links
to directory are allowed to have trailing slash,
e.g. `a.md/` is bad and `dir1/` is ok