[#150] Change `ignore link` behavior
Problem:
Currently `ignore link` annotation works for a  first link after it
(in a whole  file). That can be bad for user,
e.g. he may forgot to delete this annotation, and now it ignores link in
some random place

Solution:
Throw scan errors when `ignore link` is not followed by a node with link.
To do this, we need to increase amount  of context in `ScannerM`