[#89] Handle user interrupts
Currently, if xrefcheck is interrupted, it's not displaying any
`VerifyError`s, even if it has already found some.
When we found error, we are notifying user via progress bar, but he must
wait checking of all references to see it
(and it can be long enough).

Solution: when xrefcheck is interrupted, display all errors
we have already found and a number of
checked references.
To do this, we need to modify `forConcurrentlyCaching`
behaviour on exceptions.