[Chore] Fix dockerhub tags
Problem: We have a pipeline step to tag docker images on dockerhub
whenever a new version is released:
https://github.com/serokell/xrefcheck/blob/7dd5c4c3c954a531b5cad89857f31b27245f0ef9/.buildkite/pipeline.yml#L51-L56

However, this doesn't seem to be working, dockerhub only contains the
`latest` tag: https://hub.docker.com/r/serokell/xrefcheck/tags

The problem *seems* to be that the CI step is only triggered when it
builds a branch with a name matching the regex `/^v[0-9]+.*/`. But we
never use that format for branch names, so it's never triggered.

Solution:
1. Change the CI step to trigger when it detects a tag with a version
   number
2. Enable the "Build tags" option in buildkite:
   https://buildkite.com/serokell/xrefcheck/settings/repository