[#64] Implement copy/paste protection checks
Problem: Currently xrefcheck is not able to detect possibly bad
copy-pastes, when some links are referring the same file, but
from the link name it seems that one of
that links should refer other file.

Solution: Implement check, add support for related annotations
for `.md` files, add corresponding settings to the config.